Category and Movement validation fixed

+++ TKDprogress_BLL/Services/MovementService.cs
@@ -27,13 +27,29 @@ namespace TKDprogress_BLL.Services
 
         public async Task<MovementDto> GetMovementByIdAsync(int id)
         {
+            if (id <= 0)
+            {
+                return new MovementDto { ErrorMessage = "Invalid movement." };
+            }
+
             MovementDto movement = await _movementRepository.GetMovementByIdAsync(id);
+
+            if (movement.Id <= 0)
+            {
+                return new MovementDto { ErrorMessage = "Movement not found." };
+            }
 
             return movement;
         }
 
         public async Task<MovementDto> CreateMovementAsync(MovementDto movement)
         {
+            if (string.IsNullOrEmpty(movement.Name) || string.IsNullOrEmpty(movement.ImageUrl))
+            {
+                movement.ErrorMessage = "Movement name or image is incorrect.";
+                return movement;
+            }
+
             await _movementRepository.CreateMovementAsync(movement);
 
             return movement;
@@ -41,6 +57,18 @@ namespace TKDprogress_BLL.Services
 
         public async Task<MovementDto> UpdateMovementAsync(MovementDto movement)
         {
+            if (movement.Id <= 0)
+            {
+                movement.ErrorMessage = "Movement type does not exist.";
+                return movement;
+            }
+
+            if (string.IsNullOrEmpty(movement.Name) || string.IsNullOrEmpty(movement.ImageUrl))
+            {
+                movement.ErrorMessage = "Movement name or image is incorrect.";
+                return movement;
+            }
+
             await _movementRepository.UpdateMovementAsync(movement);
 
             return movement;
@@ -48,6 +76,12 @@ namespace TKDprogress_BLL.Services
 
         public async Task<MovementDto> DeleteMovementAsync(MovementDto movement)
         {
+            if (movement.Id <= 0)
+            {
+                movement.ErrorMessage = "Movement type does not exist.";
+                return movement;
+            }
+
             await _movementRepository.DeleteMovementAsync(movement);
 
             return movement;
